SQUASH ME

+++ parent/hsm/net.bluemind.hsm.storage/src/net/bluemind/hsm/storage/IHSMStorage.java
@@ -45,7 +45,7 @@ public interface IHSMStorage extends Closeable {
 	 * @return
 	 * @throws IOException
 	 */
-	InputStream peek(String domainUid, String mailboxUid, String hsmId, long maxMessageSize) throws IOException;
+	InputStream peek(String domainUid, String mailboxUid, String hsmId, Integer maxMessageSize) throws IOException;
 
 	/**
 	 * Delete a stored mime stream with the given HSM.
